redis: clustering fixes
Detect and deal with nodes that have joined the cluster but weren't
turned into replicas at the time. During removal we should just leave
it as a master, and during normal management it should be turned into
a replica.

Tweak loser detection to properly consider masters with no replicas. In
addition to fixing an edge case where we would choose the wrong loser,
this also fixes an infinite loop when no masters have replicas.