redis: try harder to get an initial PONG
While this is unlikely in a normal cluster, an undersized cluster can
easily have a primary with no replicas. If something interferes with the
initial PING that primary is removed without any attempt to add a
replica and fail over to it, resulting in a broken cluster.

(Undersized clusters can also just fail in normal operation, but it's
not particularly costly to retry in case there was a transient network
issue.)