python_venv: use a nonperson account to hold venvs
meson 1.1.0 breaks building some Python packages as root by trying to be
smart, detecting that it's building as root under sudo, and dropping
privileges by sudoing back to the "original" user. Since pip created the
build environment as root this results in the build not being able to
access the build environment.

These don't technically need to be owned by root, so the easiest way to
fix the build is to do it as a different user.