redis: try harder to remove nodes
If a replica fails to switch to the new primary during promotion,
waiting for the former primary to have no replicas times out. Since
this invalid topology is a situation the script is able to deal with
when it knows about it, an easy way to make this more robust is
simply to try again from the beginning.