build-ami: update base image version, stable subnet sort
I have figured out why the subnets are funky: for legacy reasons the
umcollab account has the public subnets in one VPC and the VPN subnets
in another. In order for the NAT gateway to work it needs to be on a
public subnet, so one of the unused VPN subnets was turned into a public
subnet and cannot be used as a VPN subnet.

I hate everything about this.